fix(new-client): small fixes tearout
