+++ src/new-client/src/App/TearOutSection/tearOutContext.tsx
@@ -1,0 +1,7 @@
+import { createContext } from "react"
+
+export const sectionsTornOut = {
+  isTornOut: false,
+}
+
+export const TearOutContext = createContext(sectionsTornOut)
