+++ src/new-client/src/Web/TearOutSection/handleTearOutSection.ts
@@ -1,3 +1,16 @@
-export function handleTearOutSection(section: string, close?: boolean) {
-  return new Error("Function should be implemented at platform level")
+import { constructUrl } from "@/utils/url"
+import { openWindow } from "@/utils/window/openWindow"
+import { tearOutSection, Section, sectionConfig } from "./state"
+
+export function handleTearOutSection(section: Section) {
+  const { width, height } = sectionConfig[section]
+  openWindow(
+    {
+      url: constructUrl(`/${section}`),
+      name: section,
+      width: width,
+      height: height,
+    },
+    () => tearOutSection(false, section),
+  )
 }
